Fix up format warnings, from Sam Tregar <sam@tregar.com>. (Modified to avoid conflict with Dan's changes.)
git-svn-id: https://svn.parrot.org/parrot/trunk@41 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ bytecode.c
@@ -93,7 +93,7 @@ read_constants_table(void** program_code)
         }
         num--;
         if (len < 0 || (len > 0 && num == 0)) {
-            printf("Bytecode error: string constant segment corrupted: %i, %i\n", len, num);
+            printf("Bytecode error: string constant segment corrupted: %i, %i\n", (int) len, (int) num);
             exit(1);
         }
     }
